requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.3
greenlet==0.4.10
gunicorn==19.6.0
MarkupSafe==0.23
msgpack-python==0.4.8
oauthlib==2.0.0
pdb==0.1
Pillow==3.4.2
psycopg2==2.6.2
PyJWT==1.4.2
python-dateutil==2.6.0
python-gnupg==0.3.9
python-social-auth==0.2.21
python3-openid==3.0.10
PyYAML==3.12
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
whitenoise==3.2.2
